Load file

+++ src/position.rs
@@ -1,12 +1,12 @@
 /// Real position for the map
 #[derive(Copy, Clone, Debug, Default, PartialEq)]
 pub struct Position {
-    pub x: f32,
-    pub y: f32,
+    pub x: f64,
+    pub y: f64,
 }
 
 impl Position {
-    pub fn new(x: f32, y: f32) -> Self {
+    pub fn new(x: f64, y: f64) -> Self {
         Self { x, y }
     }
 }
